make destinationPolicy optional
Updated handlers to not require destinationPolicy, aws-logs-destination.json to not list as required and added unit tests

These changes were made to resolve a github issue created by an external customer.
Customer requested destinationPolicy to not be required when creating logs destination as it is not required when creating destination via API.
Change updates the CreateHandler to not attempt action on destinationPolicy when it is not provided.
Unit tests, integration tests, and canary tests were also created to test the new edge case.