adding support for metric filter dimensions and units
This is a feature request implementation. The logs metric filters support units and dimensions through the API but this feature was not available via CFN(until now).
These changes implement the changes to allow creation of dimensions and units and also include the unit tests. The changes the to the test files are minimal since the AbstractTestBase.java file modifies the model to contain dimensions and units and thus results in all the handler test files to use the new model. To unit test handlers work without dimensions and units(since these parameters are optional), pass in false to the buildDefaultModel method.