Adding Logs MetricFilter Dimensions and Unit Support (#84)
* make destinationPolicy optional

Updated handlers to not require destinationPolicy, aws-logs-destination.json to not list as required and added unit tests

These changes were made to resolve a github issue created by an external customer.
Customer requested destinationPolicy to not be required when creating logs destination as it is not required when creating destination via API.
Change updates the CreateHandler to not attempt action on destinationPolicy when it is not provided.
Unit tests, integration tests, and canary tests were also created to test the new edge case.

* These changes are in response to Sam's PR comments.

* This commit is in response to Wai's PR comments.

* fixed typo in updateHandlerTest.java

* adding support for metric filter dimensions and units

This is a feature request implementation. The logs metric filters support units and dimensions through the API but this feature was not available via CFN(until now).
These changes implement the changes to allow creation of dimensions and units and also include the unit tests. The changes the to the test files are minimal since the AbstractTestBase.java file modifies the model to contain dimensions and units and thus results in all the handler test files to use the new model. To unit test handlers work without dimensions and units(since these parameters are optional), pass in false to the buildDefaultModel method.

* Updated UpdateHandlerTest.java to fix bug

* correct metric filter schema to match API

* updated variable name in translator.java

this commit is to accept a CR suggestion from Wai

* update AbstractTestBase.java imports

this commit is to accept a suggestion from Wai's CR

* update Translator.java imports to exclude wildcard imports

this commit is to accept suggestion from Sam's CR

* update TranslatorTest.java imports to exclude wildcard

this commit is to accept changes suggested by Sam's CR

* update Translator.java to match updated schema(.json file)

* Update AbstractTestBase.java to match schema change

* update ListHandlerTest.java to match schema change

* update TranslatorTest.java to match schema update

* update TranslatorTest.java to fix missing import bug

Co-authored-by: Muhammad Huzaifa <hzf@amazon.com>